+++ components/notifier/src/strategies/changed_status.py
@@ -4,29 +4,29 @@ from typing import Dict, List, Union
 
 
 def get_notable_metrics_from_json(
-        data_model, json, most_recent_measurement_seen: str) -> List[Dict[str, Union[str, int]]]:
+        data_model, json, most_recent_measurement_seen: str) -> List[Dict[str, Union[Dict[str, str], str]]]:
     """Return the reports that have a webhook and metrics that require notifying."""
     notifications = []
     for report in json["reports"]:
         notable_metrics = []
-        webhook = report.get("teams_webhook")
         for subject in report["subjects"].values():
             for metric in subject["metrics"].values():
                 if has_new_status(metric, "target_not_met", most_recent_measurement_seen) \
                         or has_new_status(metric, "unknown", most_recent_measurement_seen):
                     notable_metrics.append(create_notification(data_model, metric))
-        if webhook and len(notable_metrics) > 0:
+        if report.get("notification_destinations") and notable_metrics:
             notifications.append(
-                dict(report_uuid=report["report_uuid"], report_title=report["title"], teams_webhook=webhook,
-                     url=report.get("url"), metrics=notable_metrics))
+                dict(report_uuid=report["report_uuid"], report_title=report["title"],
+                     url=report.get("url"), metrics=notable_metrics,
+                     notification_destinations=report["notification_destinations"]))
     return notifications
 
 
-def create_notification(data_model, metric):
+def create_notification(data_model, metric) -> Dict[str, str]:
     """Create the notification dictionary."""
     recent_measurements = metric["recent_measurements"]
     scale = metric["scale"]
-    result = dict(
+    return dict(
         metric_type=metric["type"],
         metric_name=metric["name"] or f'{data_model["metrics"][metric["type"]]["name"]}',
         metric_unit=metric["unit"] or f'{data_model["metrics"][metric["type"]]["unit"]}',
@@ -34,7 +34,6 @@ def create_notification(data_model, metric):
         new_metric_value=recent_measurements[-1][scale]["value"],
         old_metric_status=get_status(data_model, recent_measurements[-2][scale]["status"]),
         old_metric_value=recent_measurements[-2][scale]["value"])
-    return result
 
 
 def get_status(data_model, status) -> str: