+++ components/notifier/tests/test_quality_time_notifier.py
@@ -70,6 +70,19 @@ class FakeResponse:
 class NotifyTests(unittest.IsolatedAsyncioTestCase):
     """Unit tests for the notify method."""
 
+    def setUp(self):
+        """Define info that is used in multiple tests."""
+        self.url = "https://report1"
+        self.title = "Report 1"
+        self.history = "2020-01-01T00:23:59+59:00"
+        self.subjects = dict(
+            subject1=dict(
+                metrics=dict(
+                    metric1=dict(
+                        type="tests", name="metric1", unit="units", status="target_not_met", scale="count",
+                        recent_measurements=[
+                            dict(start=self.history, end=self.history, count=dict(status="target_met", value="5"))]))))
+
     @classmethod
     def setUpClass(cls) -> None:
         """Provide the data_model to the class."""
@@ -121,16 +134,12 @@ class NotifyTests(unittest.IsolatedAsyncioTestCase):
     @patch("aiohttp.ClientSession.get")
     async def test_one_new_red_metric(self, mocked_get, mocked_sleep, mocked_send):
         """Test that a notification is not sent if there is one new red metric."""
-        history = "2020-01-01T00:23:59+59:00"
         report = dict(
-            report_uuid="report1", title="Report 1", url="https://report1", teams_webhook="https://webhook",
-            subjects=dict(
-                subject1=dict(
-                    metrics=dict(
-                        metric1=dict(
-                            type="tests", name="metric1", unit="units", status="target_not_met", scale="count",
-                            recent_measurements=[
-                                dict(start=history, end=history, count=dict(status="target_met", value="5"))])))))
+            report_uuid="report1", title=self.title, url=self.url,
+            notification_destinations=dict(
+                destination1=dict(
+                    name="destination name", teams_webhook="www.webhook.com")),
+            subjects=self.subjects)
         now = datetime.now().isoformat()
         report2 = deepcopy(report)
         report2["subjects"]["subject1"]["metrics"]["metric1"]["recent_measurements"].append(
@@ -148,9 +157,9 @@ class NotifyTests(unittest.IsolatedAsyncioTestCase):
     @patch("aiohttp.ClientSession.get")
     async def test_one_old_red_metric_with_one_measurement(self, mocked_get, mocked_sleep, mocked_send):
         """Test that a notification is not sent if there is one old red metric."""
-        history = "2020-01-01T00:23:59+59:00"
+        history = self.history
         report = dict(
-            report_uuid="report1", title="Report 1", url="https://report1", teams_webhook="https://webhook",
+            report_uuid="report1", title=self.title, url=self.url, teams_webhook="https://webhook",
             subjects=dict(
                 subject1=dict(
                     metrics=dict(
@@ -175,7 +184,31 @@ class NotifyTests(unittest.IsolatedAsyncioTestCase):
         mocked_get.side_effect = [Exception]
         mocked_sleep.side_effect = [None, RuntimeError]
         try:
-            await retrieve_data_model()
+            await retrieve_data_model("v3")
         except RuntimeError:
             pass
         self.assertEqual(2, mocked_log_warning.call_count)
+
+
+    @patch('pymsteams.connectorcard.send')
+    @patch("asyncio.sleep")
+    @patch("aiohttp.ClientSession.get")
+    async def test_no_webhook_in_notification_destination(self, mocked_get, mocked_sleep, mocked_send):
+        """Test that the notifier continues if a destination doesnt have a webhook configured."""
+        report = dict(
+            report_uuid="report1", title=self.title, url=self.url,
+            notification_destinations=dict(
+                destination1=dict(
+                    name="destination name", teams_webhook="")),
+            subjects=self.subjects)
+        now = datetime.now().isoformat()
+        report2 = deepcopy(report)
+        report2["subjects"]["subject1"]["metrics"]["metric1"]["recent_measurements"].append(
+            dict(start=now, end=now, count=dict(status="target_met", value="10")))
+        mocked_get.side_effect = [self.return_data_model(), self.return_report(report), self.return_report(report2)]
+        mocked_sleep.side_effect = [None, RuntimeError]
+        try:
+            await notify()
+        except RuntimeError:
+            pass
+        mocked_send.assert_not_called()