+++ components/server/src/server_utilities/type.py
@@ -12,6 +12,7 @@ Position = Literal["first", "last", "next", "previous"]
 
 MeasurementId = NewType("MeasurementId", str)
 MetricId = NewType("MetricId", str)
+NotificationDestinationId = NewType("NotificationDestinationId", str)
 ReportId = NewType("ReportId", str)
 SessionId = NewType("SessionId", str)
 SourceId = NewType("SourceId", str)